+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Task tray icon.
 - `Minimize the window on launched by SteamVR` setting to minimize the window.
+- `Minimize to the task tray icon` setting to hide the window.
 - `Enable Debug Logs` setting to output debug-level logs.
 
 ### Changed